+++ rules/rule_algo.py
@@ -156,6 +156,21 @@ def jailEvent(screen, player, jail, players, dices, jail_fid):
     player.jailStatus = True
     r_ui.latest_event = [f"{player.name} went to jail."]
     
+def checkBankruptcy(screen, player, bank, players, turns):
+    if player.balance < 0:
+        for field in fc.f_container:
+            if field.owner.name == player.name:
+                field.owner = bank
+                
+        players.pop((turns -1) % len(players))
+        
+    if len(players) <= 1:
+        r_ui.latest_event = [f"{players} won"]        
+        ecp = True
+    
+        while ecp:
+            ecp = r_ui.event_card(screen, True)
+    
 def eventSelector(screen, jail, players, dices, jail_fid):
     """
     Event IDs